test: 💍 split test file in two
